Added LTL model checking, renamed Atom in AtomicProposition
Implemented the algorithm by Lichtenstein and Pnueli for LTL model
checking.
Added tests for the LTL model checking.
Renamed Atom in AtomicProposition to avoid confusion with the tableu
atoms.
Fixed a bug in the DiGraph class that avoids to add an edge whose
source is not in the DiGraph.
Removed member ‘_nodes’ from DiGraph.
Added member ‘height’ in CTLS.Formula.
Improved CTL model checking tests.
Updated the example in README.md.